Move utils to images
