+++ armstrong/core/arm_content/images/presets.py
