+++ armstrong/core/arm_content/tests/__init__.py
@@ -2,9 +2,9 @@ from ._utils import TestCase, create_random_article, create_random_video
 from .arm_content_support.models import BaseContent, Article, Video
 from .admin import *
 from .fields import *
+from .images import *
 from .mixins import *
 from .models import *
-from .utils import *
 from .video import *
