+++ armstrong/core/arm_content/tests/images/__init__.py
